remove triangle
